Always use 64-bit register values
All registers are now uin64_t, even for 32-bit code, since this should be
just as efficient on 64-bit machines. Although the MIPS ISA says that
32-bit ops produce sign-extended 32-bit values in 64-bit code, it's easier
if we let the high bits be garbage instead. This does mean we have to be a
little careful whenever we read a register value though.

Functions that return in both $v0 and $v1 are now implemented in C using
the 8 byte ReturnValue struct. On macOS/Linux x86-64 and ARM64, this will
be returned in registers so there's no performance hit. On Windows x86-64
however this is too large to be returned directly.